Finished first prototype of writeback tests.
It appears there is 0 numa traffic when prefetching, but its a bit high with regular loads. Hinting that regular remote reads trigger writebacks but not with prefetch write instruction. On only 2 threads it is quite hard to see whats going on in vtune, working to extend the test to the 32 remote threads.